Move hyperlink render to shared folder
