+++ layouts/ReleasePost.tsx
@@ -1,20 +1,17 @@
+import DisqusThread from "@/components/DisqusThread";
 import MDXRenderer from "@/components/MDXRenderer";
 import MetadataHead from "@/components/MetadataHead";
 import { RawBlogContent } from "@/lib/mdx";
 import { isNone, Nullable } from "@/lib/utils";
 import Head from "next/head";
 import React from "react";
+import HyperlinkListRender, { IHyperlink } from "./shared/HyperlinkList";
+import OtherFooterRender from "./shared/OtherFooter";
 
 interface IStaffList {
     position: string;
     name: string;
     link?: string;
-}
-
-interface IHyperlink {
-    link?: string;
-    status?: string;
-    text: string;
 }
 
 interface IPostExtra {
@@ -51,51 +48,6 @@ function StaffListRender(props: { staffs?: IStaffList[]; lang?: string }) {
             <p className="my-2 text-center font-bold text-md">{lang === "id" ? "Tukang Kerja" : "Credits"}</p>
             {staffs.map((staff) => {
                 return <StaffRender key={`${staff.position}_${staff.name}`} {...staff} />;
-            })}
-        </div>
-    );
-}
-
-function HyperlinkRender({ link, text, status }: IHyperlink) {
-    if (typeof link !== "string") {
-        return (
-            <button
-                className="bg-rose-900 text-gray-200 cursor-not-allowed rounded-md p-3 text-center font-bold text-sm my-1 mx-0.5 inline-block"
-                title={status}
-            >
-                {text}
-            </button>
-        );
-    }
-
-    return (
-        <a
-            href={link}
-            className="bg-slate-500 visited:bg-slate-600 !text-white hover:!text-white rounded-md p-3 text-center font-bold text-sm my-1 mx-0.5 inline-block hover:rounded-xl no-underline hover:no-underline transition-all"
-        >
-            {text}
-        </a>
-    );
-}
-
-function HyperlinkListRender(props: { urls?: IHyperlink[]; lang?: string }) {
-    const { urls } = props;
-    if (isNone(urls)) {
-        return null;
-    }
-    if (!Array.isArray(urls)) {
-        return null;
-    }
-    if (urls.length < 1) {
-        return null;
-    }
-    const lang = props.lang || "id";
-
-    return (
-        <div id="info-download-content">
-            <p className="my-2 text-center font-bold text-md">{lang === "id" ? "Unduh" : "Download"}</p>
-            {urls.map((url, idx) => {
-                return <HyperlinkRender key={`hyperlink-${idx}_${url.text}`} {...url} />;
             })}
         </div>
     );
@@ -161,6 +113,10 @@ export default function LayoutReleasePost(props: RawBlogContent) {
                         <StaffListRender staffs={extraData?.staffLists} lang={extraData?.lang} />
                         <HyperlinkListRender urls={extraData?.urls} lang={extraData?.lang} />
                     </div>
+                    <div id="post-other" className="mb-2 text-center">
+                        <OtherFooterRender lang={extraData?.lang} />
+                    </div>
+                    {extraData.comments && <DisqusThread identifier={frontMatter.slug} />}
                 </div>
             </div>
         </>
